+++ Program.cs
@@ -3,10 +3,12 @@
 class Program
 {
 
+    static Player player = new Player();
+
     static void Main(string[] args)
     {
         Console.WriteLine("Hello, World!");
         Console.WriteLine("테스트입니다.");
-        Console.WriteLine("Branch 테스트");
+        Console.WriteLine("Branch 테스트");        
     }
 }
